Re-implemented the JEI compat
