+++ gradle.properties
@@ -14,7 +14,7 @@ minecraft_sub_version=1.18
 
 forge_version=40.1.16
 
-jei_version=9.5.5.174
+jei_version=9.4.3.122
 
 ctm_version=1.18.2-1.1.4+4
 chisel_version=+
